Moved move-deciding logic out of index.js and into topAI.js, added in the taunts
