+++ ai/Snake.js
@@ -10,16 +10,9 @@ module.exports = class Snake {
                 break;
             }
         }
-        //this.id
-        //this.name
-        //this.coords = me.coords;
+
         this.tail = [me.coords[me.coords.length - 1][0], me.coords[me.coords.length - 1][1]];;
         this.head = [me.coords[0][0], me.coords[0][1]];
-        //this.color = getRandomColor();
-        //this.taunt = getRandomTaunt();
-        ///console.log(me);
-        //  return this;
-        //return mySnake;
     }
 
     getRandomTaunt() {
@@ -27,5 +20,4 @@ module.exports = class Snake {
         var i = Math.random() * taunts.length;
         return taunts[i];
     }
-
 }
